auto generate id :

+++ src/main/java/com/example/demo/entities/RequestComponent.java
@@ -6,38 +6,46 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+import org.hibernate.annotations.GenericGenerator;
+
 @Entity
 public class RequestComponent {
-	
-	@Id 
-	@GeneratedValue(strategy=GenerationType.AUTO)
-    @Column(name="compId")
-	private int compId;
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	@Column(name = "compId")
+	private Long compId;
 	private String name;
 	private int quantity;
-	
-	public int getid() {
+
+	public Long getid() {
 		return compId;
 	}
-	public void setid(int compId) {
+
+	public void setid(Long compId) {
 		this.compId = compId;
 	}
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public int getQuantity() {
 		return quantity;
 	}
+
 	public void setQuantity(int quantity) {
 		this.quantity = quantity;
 	}
-	
-	public RequestComponent() {}
-	
-	public RequestComponent(int compId, String name, int quantity) {
+
+	public RequestComponent() {
+	}
+
+	public RequestComponent(Long compId, String name, int quantity) {
 		super();
 		this.compId = compId;
 		this.name = name;
